requirements/base.txt
Django==2.1.2
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==0.9.2
python-dateutil==2.7.4
pytz==2018.6
six==1.11.0
text-unidecode==1.2
